% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene1.R
\docType{data}
\name{gene1}
\alias{gene1}
\title{example gene 1}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 19 rows and 20 columns.
}
\usage{
data(gene1)
}
\description{
example gene 1
}
\keyword{datasets}
