% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2.R
\docType{data}
\name{gene2}
\alias{gene2}
\title{example gene 2}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 15 columns.
}
\usage{
data(gene2)
}
\description{
example gene 2
}
\keyword{datasets}
