% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta1.R
\docType{data}
\name{meta1}
\alias{meta1}
\title{example meta 1}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 19 rows and 2 columns.
}
\usage{
data(meta1)
}
\description{
example meta 1
}
\keyword{datasets}
