% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2.R
\docType{data}
\name{meta2}
\alias{meta2}
\title{example meta 2}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 7 columns.
}
\usage{
data(meta2)
}
\description{
example meta 2
}
\keyword{datasets}
