% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta3.R
\docType{data}
\name{meta3}
\alias{meta3}
\title{example meta 3}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
data(meta3)
}
\description{
example meta 3
}
\keyword{datasets}
