% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_ui}
\alias{mod_about_ui}
\alias{mod_about_server}
\title{mod_about_ui mod_about_server}
\usage{
mod_about_ui(id)

mod_about_server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. This module generates the landing page for the
application and provides a link to the github repo where users may file
an issue.
}
\details{
about UI Function
}
