% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aboutExample.R
\name{mod_aboutExample_ui}
\alias{mod_aboutExample_ui}
\title{mod_aboutExample_ui mod_aboutExample_server}
\usage{
mod_aboutExample_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. This module generates information for the 
example data that is pre-loaded when the application launches.
}
\details{
aboutExample UI Function
}
