% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aboutTreeParams.R
\name{mod_aboutTreeParams_ui}
\alias{mod_aboutTreeParams_ui}
\title{mod_aboutTreeParams_ui mod_aboutTreeParams_server}
\usage{
mod_aboutTreeParams_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module generates the landing page which
describes the tree parameters that a user can adjust.
}
\details{
aboutTreeParams Function
}
\keyword{internal}
