% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cladeFindr.R
\name{mod_cladeFindr_ui}
\alias{mod_cladeFindr_ui}
\alias{mod_cladeFindr_server}
\title{mod_cladeFindr_ui and mod_cladeFindr_server}
\usage{
mod_cladeFindr_ui(id)

mod_cladeFindr_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
