% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combineTandG.R
\name{\%>\%}
\alias{\%>\%}
\alias{mod_combineTandG_ui}
\alias{mod_combineTandG_server}
\title{mod_combineTandG_ui and mod_combineTandG_server}
\usage{
mod_combineTandG_ui(id)

mod_combineTandG_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
