% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_csvFileInput.R
\name{mod_csvFileInput_ui}
\alias{mod_csvFileInput_ui}
\alias{mod_csvFileInput_server}
\title{mod_csvFileInput_ui and mod_csvFileInput_server}
\usage{
mod_csvFileInput_ui(id, label = "CSV file")

mod_csvFileInput_server(input, output, session, stringsAsFactors)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
