% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{\%>\%}
\alias{\%>\%}
\alias{mod_dataInput_ui}
\alias{mod_dataInput_server}
\title{mod_dataInput_ui and mod_dataInput_server}
\usage{
mod_dataInput_ui(id, label)

mod_dataInput_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This modules allows one to upload csv/tsv/txt files and is combined with the displayTable module.
}
\keyword{internal}
