% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_displayTable.R
\name{mod_displayTable_ui}
\alias{mod_displayTable_ui}
\alias{mod_displayTable_server}
\title{mod_displayTable_ui and mod_displayTable_server}
\usage{
mod_displayTable_ui(id)

mod_displayTable_server(input, output, session, file)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This modules allows one to display csv/tsv/txt files in their respective upload tabs and is combined with dataInput module.
}
\keyword{internal}
