% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_displayTree.R
\name{mod_displayTree_ui}
\alias{mod_displayTree_ui}
\alias{mod_displayTree_server}
\title{displayTree UI Function}
\usage{
mod_displayTree_ui(id)

mod_displayTree_server(
  input,
  output,
  session,
  tree_file_out,
  geneObjectOutForSNP,
  align,
  tree_format,
  font,
  num_scale,
  node,
  lim,
  boot_pos,
  mid_p,
  mat_off
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module combines the tree with genetic
 distance as an S4 object that allows tree plotting and accessing the 
 combined data.
}
\details{
@title   mod_displayTree_ui and mod_displayTree_server
}
\keyword{internal}
