% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filterTable.R
\name{mod_filterTable_ui}
\alias{mod_filterTable_ui}
\alias{mod_filterTable_server}
\title{mod_filterTable_ui and mod_filterTable_server}
\usage{
mod_filterTable_ui(id)

mod_filterTable_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
