% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_package_description.R
\name{mod_package_description_ui}
\alias{mod_package_description_ui}
\alias{mod_package_description_server}
\title{mod_package_description_ui and mod_package_description_server}
\usage{
mod_package_description_ui(id)

mod_package_description_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
