% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_paramsTree.R
\name{mod_paramsTree_ui}
\alias{mod_paramsTree_ui}
\alias{mod_paramsTree_server}
\title{mod_paramsTree_ui and mod_paramsTree_server}
\usage{
mod_paramsTree_ui(id, name)

mod_paramsTree_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module holds all of the parameters that can be adjusted for the visualization of a phylogenetic tree.
}
\keyword{internal}
