% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pushButtons.R
\name{mod_pushButtons_ui}
\alias{mod_pushButtons_ui}
\alias{mod_pushButtons_server}
\title{mod_pushButtons_ui mod_pushButtons_server}
\usage{
mod_pushButtons_ui(id)

mod_pushButtons_server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. This module generates the displayed action
buttons for add/remov annotations and add/remove heatmap.
}
\details{
pushButtons UI Function
}
