% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_read_in_data.R
\name{mod_read_in_data_ui}
\alias{mod_read_in_data_ui}
\alias{mod_read_in_data_server}
\title{mod_read_in_data_ui and mod_read_in_data_server}
\usage{
mod_read_in_data_ui(id, label = " upload your meta data CSV file")

mod_read_in_data_server(input, output, session, stringsAsFactors)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
