% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_relaunchApp.R
\name{mod_relaunchApp_ui}
\alias{mod_relaunchApp_ui}
\alias{mod_relaunchApp_server}
\title{mod_relaunchApp_ui mod_relaunchApp_server}
\usage{
mod_relaunchApp_ui(id)

mod_relaunchApp_server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. This module will relaunch the application
and clear out any existing data sets.
}
\details{
relaunchApp Function
}
\keyword{internal}
