% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tipCheck.R
\name{mod_tipCheck_ui}
\alias{mod_tipCheck_ui}
\alias{mod_tipCheck_server}
\title{mod_tipCheck_ui mod_tipCheck_server}
\usage{
mod_tipCheck_ui(id)

mod_tipCheck_server(
  input,
  output,
  session,
  meta_file_out,
  m_file_out,
  gene_file_out,
  g_file_out,
  t_file_out,
  tree_file_out
)
}
\arguments{
\item{id, input, output, session}{internal}

\item{meta_file_out}{used to confirm that file is uploaded; on/off}

\item{m_file_out}{meta data that is read in used  file for tip checking}

\item{gene_file_out}{used to confirm that file is uploaded; on/off}

\item{g_file_out}{gene data that is read in used for tip checking}

\item{t_file_out}{tree data that is read in for tip checking}

\item{tree_file_out}{used to confirm that file is uploaded; on/off}
}
\description{
A shiny Module. This module generates messages for the user
regarding the files uploaded and if column(s) for a heatmap is available.
This module contains 3 functions located in the golem_utils_server
file (sanity, not_columns, and m_file_conversion). Essentially. these
functions are used for checking or converting files for downstream analyses.
}
\details{
tipCheck Function
}
