% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treeInput.R
\name{mod_treeInput_ui}
\alias{mod_treeInput_ui}
\alias{mod_treeInput_server}
\title{mod_treeInput_ui and mod_treeInput_server}
\usage{
mod_treeInput_ui(id)

mod_treeInput_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
