% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treeUpdate.R
\name{mod_treeUpdate_ui}
\alias{mod_treeUpdate_ui}
\alias{mod_treeUpdate_server}
\title{mod_treeUpdate_ui and mod_treeUpdate_server}
\usage{
mod_treeUpdate_ui(id)

mod_treeUpdate_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
