% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_updateTree.R
\name{mod_updateTree_ui}
\alias{mod_updateTree_ui}
\alias{mod_updateTree_server}
\title{mod_updateTree_ui and mod_updateTree_server}
\usage{
mod_updateTree_ui(id)

mod_updateTree_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
