% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploadTree.R
\name{mod_uploadTree_ui}
\alias{mod_uploadTree_ui}
\alias{mod_uploadTree_server}
\title{mod_uploadTree_ui and mod_uploadTree_server}
\usage{
mod_uploadTree_ui(id, label = "Upload a newick file, please")

mod_uploadTree_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module will upload a newick tree and is read in by treeio package with the function read.newick and allows one to midpoint root the tree
}
\keyword{internal}
