% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploadData_fct_readData.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(file_path, sep)
}
\arguments{
\item{file_path}{path of where to look for file}

\item{sep}{the file separator}
}
\description{
function to read in the data using readr::read_delim
filePath is the path to the location of the file you want to read in
sep is the specified delimiter, probably either a tab or comma
the other bits here help with reading in the file: trim whitespace, skip
empty row, column names, and how to read in the data; default is
set at column as characters
}
\details{
read_data Function
}
