% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(onStart = NULL, options = list(), enableBookmarking = NULL, ...)
}
\arguments{
\item{...}{other arguments}
}
\description{
Run the Shiny Application
}
