% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree1.R
\docType{data}
\name{tree1}
\alias{tree1}
\title{Example Tree 1

This is a Newick phylogenetic tree for 19 isolates of Salmonella enterica
to be used with the optional genetic distance matrix (example gene 1) and
 the optional meta data file (example meta 1).}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(tree1)
}
\description{
Example Tree 1

This is a Newick phylogenetic tree for 19 isolates of Salmonella enterica
to be used with the optional genetic distance matrix (example gene 1) and
 the optional meta data file (example meta 1).
}
\examples{
data(tree1)
treeout <- ggtree::ggtree(tree1)
treeout <- treeout + ggtree::geom_tiplab() #add labels to the tree
}
\keyword{datasets}
