% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree1504MLEXH.data.R
\docType{data}
\name{tree1504MLEXH_1}
\alias{tree1504MLEXH_1}
\title{Newick formatted phylogenetic tree}
\format{
An object of class
}
\usage{
data(tree1504MLEXH_1)
}
\description{
Data - need more of a description
}
\keyword{datasets}
