% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2.R
\docType{data}
\name{tree2}
\alias{tree2}
\title{Example Tree 2

#' This is a Newick phylogenetic tree for 19 isolates of Eschericia coli
to be used with the optional genetic distance matrix (example gene 2) and
the optional meta data file (example meta 2).}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(tree2)
}
\description{
Example Tree 2

#' This is a Newick phylogenetic tree for 19 isolates of Eschericia coli
to be used with the optional genetic distance matrix (example gene 2) and
the optional meta data file (example meta 2).
}
\examples{
data(tree2)
tree2 <- ggtree::ggtree(tree2)
}
\keyword{datasets}
