% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree3.R
\docType{data}
\name{tree3}
\alias{tree3}
\title{Example Tree 3

This is a Newick phylogenetic tree for 19 isolates of Salmonella enterica sub
species enterica to be used with the genetic distance matrix
(example gene 3) and the optional meta data file (example meta 3).}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(tree3)
}
\description{
Example Tree 3

This is a Newick phylogenetic tree for 19 isolates of Salmonella enterica sub
species enterica to be used with the genetic distance matrix
(example gene 3) and the optional meta data file (example meta 3).
}
\examples{
data(tree3)
tree3 <- ape::plot.phylo(tree3)
}
\keyword{datasets}
