% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene3.R
\docType{data}
\name{gene3}
\alias{gene3}
\title{example gene 3}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 17 columns.
}
\usage{
data(gene3)
}
\description{
example gene 3
}
\keyword{datasets}
