% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2.R
\docType{data}
\name{meta2}
\alias{meta2}
\title{Example Meta 2

This is a meta data file for tip correction for 19 isolates of Salmonella
enterica sub species enterica to be used with the newick phylogenetic tree
(example tree 2) and the optional genetic distance file (example gene 2).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 4 columns.
}
\usage{
data(meta2)
}
\description{
Example Meta 2

This is a meta data file for tip correction for 19 isolates of Salmonella
enterica sub species enterica to be used with the newick phylogenetic tree
(example tree 2) and the optional genetic distance file (example gene 2).
}
\examples{
data(meta2)
 #convert the column Display labels to the row name
meta2 <- meta2 \%>\% tibble::column_to_rownames(var = "Display.labels")\%>\%
dplyr::select(-Tip.labels) #exclude column Tip.labels
}
\keyword{datasets}
