% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aboutExample.R
\name{mod_aboutExample_ui}
\alias{mod_aboutExample_ui}
\title{mod_aboutExample_ui mod_aboutExample_server}
\usage{
mod_aboutExample_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module generates the tab, which 
contains information regarding the example data sets that is pre-loaded
 when the application launches.
}
\details{
aboutExample Function
}
\keyword{internal}
