% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_displayTree.R
\name{mod_displayTree_ui}
\alias{mod_displayTree_ui}
\alias{mod_displayTree_server}
\title{mod_displayTree_ui and mod_displayTree_server}
\usage{
mod_displayTree_ui(id)

mod_displayTree_server(input, output, session, outTree)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
