% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadImage.R
\name{mod_downloadImage_ui}
\alias{mod_downloadImage_ui}
\alias{mod_downloadImage_server}
\title{mod_downloadImage_ui and mod_downloadImage_server}
\usage{
mod_downloadImage_ui(id)

mod_downloadImage_server(input, output, session, tree_out)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{treeout}{from cladeAnnotator module; mostly like class of ggtree, gg, 
gglot2 if user combined genetic data with tree data; if not then class phylo}
}
\description{
A shiny Module. This module allows one to download the
adjusted tree image. Current download formats include png, pdf, and tiff.
}
\details{
downloadImage Function
}
\keyword{internal}
