% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadImage.R
\name{mod_downloadImage_ui}
\alias{mod_downloadImage_ui}
\alias{mod_downloadImage_server}
\title{mod_downloadImage_ui and mod_downloadImage_server}
\usage{
mod_downloadImage_ui(id, label)

mod_downloadImage_server(input, output, session, treeWLayers)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module allows one to download the adjusted tree image. Current download formats include png, pdf, and tiff.
}
\keyword{internal}
