% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exampleData.R
\name{mod_exampleData_ui}
\alias{mod_exampleData_ui}
\alias{mod_exampleData_server}
\title{mod_exampleData_ui and mod_exampleData_server}
\usage{
mod_exampleData_ui(id)

mod_exampleData_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module sources the pre-loaded example data (e.g. tree (x2), genetic distance file (x2), and meta data file (x2))
}
\keyword{internal}
