% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exampleData.R
\name{mod_exampleData_ui}
\alias{mod_exampleData_ui}
\alias{mod_exampleData_server}
\title{mod_exampleData_ui and mod_exampleData_server}
\usage{
mod_exampleData_ui(id)

mod_exampleData_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. This module sources the pre-loaded
example data (e.g. tree (x3), genetic distance file (x3), and meta data file
(x3)), which are already combined as a dataset (e.g. example data 1 is the
combined data files of example tree 1, example genetic 1, and example 
meta 1.) This module contains 3 functions located in the golem_utils_server
file (replace_column_header, gene_object_out, m_file_conversion). 
Essentially, these functions are used to manipulate the files to combine
them.
}
\details{
exampleData function
}
\keyword{internal}
