% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_FAQ.R
\name{mod_FAQ_ui}
\alias{mod_FAQ_ui}
\title{mod_about_ui mod_about_server}
\usage{
mod_FAQ_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. This module generates the FAQ page for the
application, which provides users with commmon tips for issues.
}
\details{
FAQ UI Function
}
