% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tipCheck.R
\name{mod_tipCheck_ui}
\alias{mod_tipCheck_ui}
\alias{mod_tipCheck_server}
\title{mod_tipCheck_ui mod_tipCheck_server}
\usage{
mod_tipCheck_ui(id)

mod_tipCheck_server(
  input,
  output,
  session,
  mFileOut,
  metaFileOut,
  gFileOut,
  tFileOut
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. This module generates the landing page for the 
application and provides a link to the github repo where users may file
an issue
}
\details{
tipCheck UI Function
}
