% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tipCheck.R
\name{mod_tipCheck_ui}
\alias{mod_tipCheck_ui}
\alias{mod_tipCheck_server}
\title{mod_tipCheck_ui mod_tipCheck_server}
\usage{
mod_tipCheck_ui(id)

mod_tipCheck_server(
  input,
  output,
  session,
  meta_file_out,
  m_file_out,
  gene_file_out,
  g_file_out,
  t_file_out,
  tree_file_out
)
}
\arguments{
\item{id, input, output, session}{internal}

\item{meta_file_out}{imported meta file for confirming that file is uploaded}

\item{m_file_out}{imported meta data file for tip checking}

\item{gene_file_out}{imported gene file for confirming that file is uploaded}

\item{g_file_out}{imported genetic distance file for tip checking}

\item{t_file_out}{imported tree for tip checking}
}
\description{
A shiny Module. This module generates the landing page for the
application and provides a link to the github repo where users may file
an issue
}
\details{
tipCheck UI Function
}
