% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinselR.R
\docType{package}
\name{tinselR}
\alias{tinselR}
\title{tinselR: An R Shiny Application for annotating outbreak phylogenetic trees}
\description{
This is a phylogenetic tree application for annotating trees with additional
information. Additional information could come from a genetic distance file or
meta data file.
}
\section{tinselR modules}{

The main modules of tinselR include:
mod_about -
}

