% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree3.R
\docType{data}
\name{tree3}
\alias{tree3}
\title{This is a newick phylogenetic tree for 19 isolates of Salmonella enterica
to be used with the optional genetic distance matrix (example gene 3) and the 
optional meta data file (example meta 3).}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(tree3)
}
\description{
This is a newick phylogenetic tree for 19 isolates of Salmonella enterica
to be used with the optional genetic distance matrix (example gene 3) and the 
optional meta data file (example meta 3).
}
\examples{
data(tree3)
treeout <- ggtree::ggtree(tree3) 
treeout <- treeout + ggtree::geom_tiplab() #add labels to the tree
}
\keyword{datasets}
